dashboard user & admin page created

+++ src/Pages/Dashboard/MyOrders/MyOrders.js
@@ -18,21 +18,23 @@ const MyOrders = () => {
   const { email } = user;
   useEffect(() => {
     axios.get(`http://localhost:5000/orders?email=${email}`).then(res => {
-      console.log(res);
+      // console.log(res);
       setOrders(res.data);
     });
   }, [email]);
 
   const handleDeleteOrder = id => {
-    console.log(id);
-    axios.delete(`http://localhost:5000/orders/${id}`).then(res => {
-      // console.log(data);
-      if (res.data.deletedCount === 1) {
-        alert('Successfully Deleted Service');
-        const remainingOrders = orders.filter(order => order._id !== id);
-        setOrders(remainingOrders);
-      }
-    });
+    const sure = window.confirm('Are You Sure to Delete this Order?');
+    if (sure) {
+      axios.delete(`http://localhost:5000/orders/${id}`).then(res => {
+        // console.log(data);
+        if (res.data.deletedCount === 1) {
+          alert('Successfully Deleted Service');
+          const remainingOrders = orders.filter(order => order._id !== id);
+          setOrders(remainingOrders);
+        }
+      });
+    }
   };
   return (
     <div>
